+++ src/argilla_server/policies.py
@@ -77,6 +77,17 @@ class WorkspaceUserPolicy:
         return is_allowed
 
 
+class WorkspaceUserPolicyV1:
+    @classmethod
+    def list(cls, workspace_id: UUID) -> PolicyAction:
+        async def is_allowed(actor: User) -> bool:
+            return actor.is_owner or (
+                actor.is_admin and await _exists_workspace_user_by_user_and_workspace_id(actor, workspace_id)
+            )
+
+        return is_allowed
+
+
 class WorkspacePolicy:
     @classmethod
     async def list(cls, actor: User) -> bool: